Set subsystem to 5.01 (XP)
I was originally going to do 5.00 but apparently the Microsoft linker is a bitch and will override invalid subsystem types. This also means 64 bit builds will ignore 5.01 and default to 6.00, but I'm too fucking retard to figure out how to port the XP target to 5.15 so guess it'll stay fucked. Only building x86 for XP anyways.